lec 21:Adding Authentication to Journal Endpoints

+++ src/main/java/com/vinayak/journalApp/service/JournalEntryService.java
@@ -26,13 +26,16 @@ public class JournalEntryService {
     public void saveEntry(JournalEntry journalEntry, String userName) {
         try {
             User user = userService.findByUserName(userName);
+            if (user.getJournalEntries() == null) {
+                user.setJournalEntries(new ArrayList<>());
+            }
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e){
             System.out.println(e);
-            throw new RuntimeException("Error saving JournalEntry");
+            throw new RuntimeException("Error saving JournalEntry",e);
         }
     }
     public void saveEntry(JournalEntry journalEntry) {
@@ -47,10 +50,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed=false;
+        try{
         User user= userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        removed =user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed){
+            userService.saveUser(user);
+            journalEntryRepository.deleteById(id);
+        }
+    }catch(Exception e)
+        {
+            System.out.println(e);
+            throw new RuntimeException("Error deleting JournalEntry",e);
+        }
+        return removed;
     }
 }
